requirements.txt
langchain
langchain-community
boto3
python-dotenv
loguru
PyYAML